+++ services/dcs/v2/model/CreateCustomTemplateRequest.ts
@@ -1,0 +1,12 @@
+import { CreateCustomTemplateBody } from './CreateCustomTemplateBody';
+
+
+export class CreateCustomTemplateRequest {
+    public body?: CreateCustomTemplateBody;
+    public constructor() { 
+    }
+    public withBody(body: CreateCustomTemplateBody): CreateCustomTemplateRequest {
+        this['body'] = body;
+        return this;
+    }
+}