+++ services/dcs/v2/model/CreateCustomTemplateResponse.ts
@@ -1,0 +1,13 @@
+
+import { SdkResponse } from "@huaweicloud/huaweicloud-sdk-core/SdkResponse";
+
+export class CreateCustomTemplateResponse extends SdkResponse {
+    public id?: string;
+    public constructor() { 
+        super();
+    }
+    public withId(id: string): CreateCustomTemplateResponse {
+        this['id'] = id;
+        return this;
+    }
+}