+++ services/frs/v2/model/LiveDetectBase64Req.ts
@@ -1,22 +1,13 @@
 
 
 export class LiveDetectBase64Req {
-    private 'action_time'?: string | undefined;
     private 'video_base64': string | undefined;
     public actions: string;
+    private 'action_time'?: string | undefined;
+    private 'nod_threshold'?: number | undefined;
     public constructor(videoBase64?: any, actions?: any) { 
         this['video_base64'] = videoBase64;
         this['actions'] = actions;
-    }
-    public withActionTime(actionTime: string): LiveDetectBase64Req {
-        this['action_time'] = actionTime;
-        return this;
-    }
-    public set actionTime(actionTime: string | undefined) {
-        this['action_time'] = actionTime;
-    }
-    public get actionTime() {
-        return this['action_time'];
     }
     public withVideoBase64(videoBase64: string): LiveDetectBase64Req {
         this['video_base64'] = videoBase64;
@@ -31,5 +22,25 @@ export class LiveDetectBase64Req {
     public withActions(actions: string): LiveDetectBase64Req {
         this['actions'] = actions;
         return this;
+    }
+    public withActionTime(actionTime: string): LiveDetectBase64Req {
+        this['action_time'] = actionTime;
+        return this;
+    }
+    public set actionTime(actionTime: string | undefined) {
+        this['action_time'] = actionTime;
+    }
+    public get actionTime() {
+        return this['action_time'];
+    }
+    public withNodThreshold(nodThreshold: number): LiveDetectBase64Req {
+        this['nod_threshold'] = nodThreshold;
+        return this;
+    }
+    public set nodThreshold(nodThreshold: number | undefined) {
+        this['nod_threshold'] = nodThreshold;
+    }
+    public get nodThreshold() {
+        return this['nod_threshold'];
     }
 }