+++ services/frs/v2/model/LiveDetectUrlReq.ts
@@ -1,22 +1,13 @@
 
 
 export class LiveDetectUrlReq {
-    private 'action_time'?: string | undefined;
     private 'video_url': string | undefined;
     public actions: string;
+    private 'action_time'?: string | undefined;
+    private 'nod_threshold'?: number | undefined;
     public constructor(videoUrl?: any, actions?: any) { 
         this['video_url'] = videoUrl;
         this['actions'] = actions;
-    }
-    public withActionTime(actionTime: string): LiveDetectUrlReq {
-        this['action_time'] = actionTime;
-        return this;
-    }
-    public set actionTime(actionTime: string | undefined) {
-        this['action_time'] = actionTime;
-    }
-    public get actionTime() {
-        return this['action_time'];
     }
     public withVideoUrl(videoUrl: string): LiveDetectUrlReq {
         this['video_url'] = videoUrl;
@@ -31,5 +22,25 @@ export class LiveDetectUrlReq {
     public withActions(actions: string): LiveDetectUrlReq {
         this['actions'] = actions;
         return this;
+    }
+    public withActionTime(actionTime: string): LiveDetectUrlReq {
+        this['action_time'] = actionTime;
+        return this;
+    }
+    public set actionTime(actionTime: string | undefined) {
+        this['action_time'] = actionTime;
+    }
+    public get actionTime() {
+        return this['action_time'];
+    }
+    public withNodThreshold(nodThreshold: number): LiveDetectUrlReq {
+        this['nod_threshold'] = nodThreshold;
+        return this;
+    }
+    public set nodThreshold(nodThreshold: number | undefined) {
+        this['nod_threshold'] = nodThreshold;
+    }
+    public get nodThreshold() {
+        return this['nod_threshold'];
     }
 }