+++ services/image/v2/model/RecaptureDetectResponseResultDetail.ts
@@ -1,0 +1,16 @@
+
+
+export class RecaptureDetectResponseResultDetail {
+    public label?: string;
+    public confidence?: string;
+    public constructor() { 
+    }
+    public withLabel(label: string): RecaptureDetectResponseResultDetail {
+        this['label'] = label;
+        return this;
+    }
+    public withConfidence(confidence: string): RecaptureDetectResponseResultDetail {
+        this['confidence'] = confidence;
+        return this;
+    }
+}