+++ services/image/v2/model/RunImageSuperResolutionRequest.ts
@@ -1,0 +1,12 @@
+import { ImageSuperResolutionReq } from './ImageSuperResolutionReq';
+
+
+export class RunImageSuperResolutionRequest {
+    public body?: ImageSuperResolutionReq;
+    public constructor() { 
+    }
+    public withBody(body: ImageSuperResolutionReq): RunImageSuperResolutionRequest {
+        this['body'] = body;
+        return this;
+    }
+}