+++ services/image/v2/model/RunRecaptureDetectRequest.ts
@@ -1,0 +1,12 @@
+import { RecaptureDetectReq } from './RecaptureDetectReq';
+
+
+export class RunRecaptureDetectRequest {
+    public body?: RecaptureDetectReq;
+    public constructor() { 
+    }
+    public withBody(body: RecaptureDetectReq): RunRecaptureDetectRequest {
+        this['body'] = body;
+        return this;
+    }
+}