+++ services/waf/v1/model/UpdateHostRequestBody.ts
@@ -157,8 +157,7 @@ export class UpdateHostRequestBody {
 export enum UpdateHostRequestBodyTlsEnum {
     TLS_V1_0 = 'TLS v1.0',
     TLS_V1_1 = 'TLS v1.1',
-    TLS_V1_2 = 'TLS v1.2',
-    TLS_V1_3 = 'TLS v1.3'
+    TLS_V1_2 = 'TLS v1.2'
 }
 /**
     * @export